backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.25.2
python-dotenv==1.0.0
jobspy==1.1.76
python-multipart==0.0.6
# PDF generation dependencies
reportlab==4.0.7
weasyprint==61.2